requirements.txt
httpx==0.23.0
PyYAML==6.0
pyinstaller==5.1